* Source/art/blit.m(MPRE): Remove compiler warning.         * Source/art/image.m(-DPSimage:::::::::::): Correct parameter         types. Fixes bug #41274.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@37615 72102866-910b-0410-8b05-ffd578937521

+++ Source/art/image.m
@@ -606,9 +606,9 @@ seem to cause edges to be off by a pixel
 
 
 - (void)DPSimage: (NSAffineTransform *) matrix
-                : (int) pixelsWide : (int) pixelsHigh
-                : (int) bitsPerSample : (int) samplesPerPixel 
-                : (int) bitsPerPixel : (int) bytesPerRow : (BOOL) isPlanar
+                : (NSInteger) pixelsWide : (NSInteger) pixelsHigh
+                : (NSInteger) bitsPerSample : (NSInteger) samplesPerPixel 
+                : (NSInteger) bitsPerPixel : (NSInteger) bytesPerRow : (BOOL) isPlanar
                 : (BOOL) hasAlpha : (NSString *) colorSpaceName
                 : (const unsigned char *const [5]) data
 {
